move `wait` module into `sync`
