+++ maitake/src/lib.rs
@@ -17,4 +17,3 @@ pub mod future;
 pub mod scheduler;
 pub mod sync;
 pub mod task;
-pub mod wait;
