--- maitake/src/sync/tests.rs
@@ -1,2 +1,0 @@
-#[cfg(loom)]
-mod loom_mutex;
